General functions in a separate file
Maybe individual tasks in separate files, seems more organised to me.
