+++ README
@@ -1,0 +1,12 @@
+Project contributors:
+Frank Otto
+Bauke Jansen
+Ravi van de Berg
+Olaf Massen
+Geert Schulpen
+
+
+
+
+Project setup:
+The quantummodule.py file consists of the core functions used in this project blah blah meer gepraat lijkt zo professioneel.
